ui/pyproject.toml
[tool.poetry]
name = "trustix-dash"
version = "0.1.0"
description = ""
authors = ["adisbladis <adam.hose@tweag.io>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.4"
ijson = "^3.1.3"
pygithub = "^1.54"
python-dateutil = "^2.8.1"
grpcio = "^1.34.0"

[tool.poetry.dev-dependencies]
mypy = "^0.790"
black = "^20.8b1"
django-stubs = "^1.7.0"
flake8 = "^3.8.4"
pytest = "^6.1.2"
ipython = "^7.19.0"
grpcio-tools = "^1.34.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
